Better link to .blfs_profile; PHP,FLTK,QScintilla
Also Apache, a murgalua patch (no script yet but the
patch is a biggie!); tidyHTML5;

Moved some of the config files; did a little messing
with the vim stuff.

Quite a few updates to the skeleton script I believe; one
that I sort of backtracked on was managing multiple versions
of the same program: intended to put libs and headers in
versioned subfolders and link the desired ones from the
root lib/include folder; however, there was still the
issue of bin and man files that would be trickier to
version; plus the fact that libs and includes are often
dependent on the search path order so even though the desired
version is properly linked the other version may be
found first! Something to work on though; I just didn't feel
like putting in the time to do it right. See FLTK, I may
have left the code there.